refactor: add fixes after review
