--- src/components/MainPageRoute/index.js
@@ -1,3 +1,0 @@
-import MainPageRoute from './MainPageRoute';
-
-export default MainPageRoute;
