+++ src/components/PublicRoute/index.js
@@ -1,0 +1,3 @@
+import PublicRoute from './PublicRoute';
+
+export default PublicRoute;
